memory use Uint8Array underneath the hood
There is a separate array of memory for each of the four segments in
user space and a fifth array for all of kernel (negative) space.

Begin refactoring all operations to use pure functions.  They take
machine state as the first argument and generate a diff object rather
than attaching to the prototype and accessing and mutating state
through the this reference.