subroutine linkage
Implement POP and add full program test calling a subroutine.
Add tau getter to instances of MMIX to support POP instruction.
Add special case to applyDiff utility for POP instructions, because the general
register set in the diff in a POP is for the hole, which will only be accessible
to set after rO has descreased. From the subroutine the hole is $-1.